Add V2.0

+++ script.py
@@ -2,8 +2,8 @@ import os
 import numpy as numpy
 
 def main(): 
-    a = float(os.environ["a"])
-    b= float(os.environ["b"])
+    a = int(os.environ["a"])
+    b = int(os.environ["b"])
     print("Affiche A * B : " (a*b))
     # valeurAB = float(os.environ["A"]) * float(os.environ["B"])
     # print("Affiche A * B : " + str(valeurAB))
